pypi: setup.py to enable deploy to pypi
setup.py will allow to deploy DeOldify to
PyPI and make it pip installable.

This patch itself does not deploy to PyPI,
a configuration file is required to deploy
automatically when there's a new release.

For now, it does not include the README into
the description because it will be to hard to
understand on PyPI. Another patch will be
sent to create a description that can
actually fit PyPI.